The DragDrop plugin should disable toolbars when dragging.
