+++ packages/ckeditor5-widget/src/widgettoolbarrepository.js
@@ -100,15 +100,6 @@ export default class WidgetToolbarRepository extends Plugin {
 		}, { priority: 'low' } );
 	}
 
-	/**
-	 * @inheritDoc
-	 */
-	afterInit() {
-		if ( this.editor.plugins.has( 'DragDrop' ) ) {
-			this._disableToolbarsWhenDraggingWidgets();
-		}
-	}
-
 	destroy() {
 		super.destroy();
 
@@ -275,37 +266,6 @@ export default class WidgetToolbarRepository extends Plugin {
 	 */
 	_isToolbarInBalloon( toolbar ) {
 		return this._balloon.hasView( toolbar.view );
-	}
-
-	/**
-	 * Disable the `WidgetToolbarRepository` plugin whenever the widget is being dragged into hiding all toolbars.
-	 * That can be problematic if the drag should end in the place covered by the toolbar view.
-	 *
-	 * @private
-	 */
-	_disableToolbarsWhenDraggingWidgets() {
-		const editor = this.editor;
-		const viewDocument = editor.editing.view.document;
-		const dragDrop = editor.plugins.get( 'DragDrop' );
-
-		// When starting dragging the widget, do not display any toolbar. See: #9566.
-		this.listenTo( viewDocument, 'dragstart', ( evt, data ) => {
-			const widget = data.target ? dragDrop.findDraggableWidget( data.target ) : null;
-
-			if ( widget ) {
-				this.forceDisabled( 'WidgetToolbar:drag&drop' );
-			}
-		} );
-
-		// Enable the plugin after dropping the widget in the editor...
-		this.listenTo( viewDocument, 'drop', () => {
-			this.clearForceDisabled( 'WidgetToolbar:drag&drop' );
-		} );
-
-		// ...or if drag ends outside the editable area.
-		this.listenTo( viewDocument, 'dragend', () => {
-			this.clearForceDisabled( 'WidgetToolbar:drag&drop' );
-		} );
 	}
 }
